// backend/src/domain/models/Education.ts
export interface EducationData {
  id?: number;
  institution: string;
  title: string;
  startDate: string | Date;
  endDate?: string | Date | null;
  candidateId?: number;
}

export class Education {
  id?: number;
  institution: string;
  title: string;
  startDate: Date;
  endDate?: Date;
  candidateId?: number;

  constructor(data: EducationData) {
    this.id = data.id;
    this.institution = data.institution;
    this.title = data.title;
    this.startDate = new Date(data.startDate);
    this.endDate = data.endDate ? new Date(data.endDate) : undefined;
    this.candidateId = data.candidateId;
  }
}